Switch to using non-3rd-party-malleable transaction ids
Summary: Since we discussed earlier the plan to use discretely named "Stripped" hashes, using the stripped versions in various places made for a lot of renaming. Seeing the number of renames might suggest just redefining the transaction id as the stripped one and making a new name for the non-stripped. I also had to workaround an issue where the Coinbase transaction id was not unique, to solve that I added an OP_RETURN output with the next block height.

Reviewers: ben, johann

Reviewed By: ben, johann

Differential Revision: https://phabricator.bitgo.com/D4115