blockchain: Increase the reorganization safety level to match the increased block timing, this prevents a reorganization deadlock when more than 2 hours have passed
Summary: After investigating what was causing a reorganization failure, it was tracked back to a constant that limits the number of orphan blocks. Changing this constant cannot be done after the fact, wiping chainstate is required to resolve the deadlock situation. To prevent this from happening, the constant is increased in proportion to the increased block time.

Reviewers: johann, ben

Reviewed By: ben

Differential Revision: https://phabricator.bitgo.com/D4183