Add support for RPCHash config/cmdline param which allows not specifying password
Summary:
- CME GIS has requested that the node be able to run with an RPC server configured
that does not use a password specified in a disk-based file.

- To that end, the --rpchash and --rpclimithash cmdline args (& config options)
are added. The value used for these args is the SHA256 of the HTTP Basic Auth header,
just as is stored and used internally.

- If a new config file is generated from the sample config file, prova will
automatically generate a random username and password and the corresponding
SHA256 hash in the new config file. The username and password may be
subsequently removed by the user after having saved them somewhere secure.

- In order for the provactl command to interact with the RPC server, it must
know the username + password. If this cannot placed in a config file for whatever
reason, the only other existing option is to use a command-line argument. This
is far worse than using a owner-only file from a security perspective, since
command-line arguments are visible to any user with access to the machine, via
the process list. To solve this problem, if no password is provided to the
provactl command via command-line or config file, a password is prompted from
the user via stdin.

Reviewers: benchan, alexbosworth

Reviewed By: alexbosworth

Differential Revision: https://phabricator.bitgo.com/D5701