peer: Increase trickle timeout to prevent peer networking getting stuck
Summary:
Fixes a scenario found during running large numbers of nodes and generators in a low difficulty setting where peers stop progressing their chain and need to cycle their peers a couple times to get to the chain head

The issue stems from a feedback loop with PushGetBlocksMsg and orphan blocks. The peer logic gets stuck in a loop of PushGetBlocks and then filtering out the request as a duplicate.

There is a scenario where a peer will receive an inv for an orphan block that it already has. This happens because when the orphan block was originally received it triggered requests for the parent blocks, and then the number of blocks missing were too many to send in a single message. After sending the missing blocks the remote peer then sends the original orphan block in an inv to continue the cycle of fetching parents. When this happens too quickly the peer can get stuck wanting to request an orphan block but also seeing it as a duplicate .

There may be a way to better coordinate the async process of these back and forth messages so that a deadlock is not possible, but I didn't see an obvious way to do that.

A test that reproduces this failure fairly reliably is "fast network start".

Reviewers: jameson, ben

Reviewed By: jameson, ben

Differential Revision: https://phabricator.bitgo.com/D5742